+++ app/vtselect/traces/tracecommon/tracecommon.go
@@ -24,14 +24,16 @@ var (
 	TraceSearchStep = flag.Duration("search.traceSearchStep", 24*time.Hour, "Splits the [0, now] time range into many small time ranges by -search.traceSearchStep "+
 		"when searching for spans by trace_id. Once it finds spans in a time range, it performs an additional search according to -search.traceMaxDurationWindow and then stops. "+
 		"It affects Jaeger's /api/traces/<trace_id> API.")
-	TraceMaxServiceNameList = flag.Uint64("search.traceMaxServiceNameList", 1000, "The maximum number of service name can return in a get service name request. "+
-		"This limit affects Jaeger's /api/services API.")
-	TraceMaxSpanNameList = flag.Uint64("search.traceMaxSpanNameList", 1000, "The maximum number of span name can return in a get span name request. "+
-		"This limit affects Jaeger's /api/services/*/operations API.")
-	TraceMaxTraces = flag.Int64("search.maxTraces", 1000, "The maximum number of traces/tags that can be returned in a single search request. "+
-		"Users may request with different limit value via query argument which shouldn't exceed this limit. This limit applies to Jaeger’s /api/traces API and all Tempo query APIs.")
+	TraceMaxTraces = flag.Uint64("search.maxTraces", 1000, "The maximum number of traces that can be returned in a single search request. "+
+		"Users may request with different limit value via query argument which shouldn't exceed this limit. This limit applies to Jaeger’s /api/traces API and Tempo's /api/search API.")
+	TraceMaxTags = flag.Uint64("search.maxTags", 1000, "The maximum number of tags (including service name, span name) that can be returned in a single search request. "+
+		"This limit applies to Jaeger’s /api/services, /api/services/*/operations APIs, and various Tempo tag-related APIs.")
 
 	LatencyOffset = flag.Duration("search.latencyOffset", 30*time.Second, "The time when a trace become visible in query results after the collection. see -insert.traceMaxDuration as well. (default 30s)")
+
+	// deprecated flags. preserve here for backward compatibility. should be removed in future version.
+	_ = flag.Uint64("search.traceMaxServiceNameList", 1000, "Deprecated, see -search.maxTags.")
+	_ = flag.Uint64("search.traceMaxSpanNameList", 1000, "Deprecated, see -search.maxTags.")
 )
 
 var (